requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
colorama==0.4.6
davey==0.1.5
discord.py==2.7.1
frozenlist==1.8.0
idna==3.11
iniconfig==2.3.0
multidict==6.7.1
packaging==26.0
pluggy==1.6.0
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
PyNaCl==1.5.0
pytest==9.0.3
pytest-asyncio==1.3.0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.33.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
yt-dlp==2026.3.17
